==> personal_project/0_related_works.ipynb <==
# 나무위키 & 위키피디아로부터 사용자사전 추가하기
* 데이터베이스 다운로드할 수 있는 곳
> * [나무위키](https://mu-star.net/wikidb)
> * [위키피디아](https://dumps.wikimedia.org/kowiki/)

## 나무위키의 제목 추출

import json
import sys
filename = 'C:/Users/pc/Documents/GitHub/Private_Project/personal_project/data/wikidumps/namuwiki_20170327.json'
with open(filename,'r',encoding='utf-8') as data_file:
    data = json.load(data_file )
    data_file.close()
with open('./data/out_namu_wiki2.txt','w',encoding='utf-8') as f:
    for article in data:
        if 'title' in article.keys():
            f.write(article['title']+'\n')

## 위키피디아의 제목 추출
* Python Package [wikiextractor](https://github.com/attardi/wikiextractor)
> !python WikiExtractor.py < WikiPedia dump file >

[이후 과정-1](https://github.com/buenhygoon/Private_Project/blob/master/personal_project/1_Making_User_Dictionary_1.ipynb)  
[이후 과정-2](https://github.com/buenhygoon/Private_Project/blob/master/personal_project/1_Making_User_Dictionary_2.ipynb)

# 자연어처리를 위해 주로 사용한 파이썬 패키지
!pip install gensim
!pip install customized_konlpy

* Customized Konlpy
> 파이썬 라이브러리가 설치된 폴더에 만들어진 텍스트 파일을 옮겨 놓으면 된다.   
> 맥의 경우, /Users/hyunyoun/anaconda3/lib/python3.6/site-packages/ckonlpy/data/twitter/

* Mecab
> [Mecab-ko](https://bitbucket.org/eunjeon/mecab-ko)
> [Mecab-ko-dic](https://bitbucket.org/eunjeon/mecab-ko-dic)
> [Mecab python module](https://bitbucket.org/eunjeon/mecab-python-0.996)

> Mecab의 윈도우 설치 및 사용자사전 추가는 https://github.com/Pusnow/mecab-ko-dic-msvc 를 참고  
> ~/mecab-ko-dic/user-dic/ 에 만들어진 텍스트파일을 넣어두고 ~/tools/add-userdic.sh 를 실행
