==> personal_project/1_mongodb_duplicate_row_remove.ipynb <==
import pymongo
import pandas as pd

def ToMongoDB(username, password, dbPath):
    connection = pymongo.MongoClient('mongodb://{}:{}@{}'.format(username, password, dbPath))
    return connection

def AWS_MongoDB_Information():
    import pickle
    file = pickle.load(open('./aws_mongo_User_Information','rb'))
    user = file['user']
    password = file['password']
    webpath = file['webpath']
    return user, password, webpath

def Use_Database(connection, dbname):
    db = connection[dbname]
    return db

def Use_Collection(db, collectionName):
    collection = db[collectionName]
    return collection

def Close_db(db):
    db.close()
if __name__ == "__main__":
    db_name = 'hy_db'
    mongodb = ToMongoDB(*AWS_MongoDB_Information())
    dbname = db_name
    useDb = Use_Database(mongodb, dbname)

useCollection = Use_Collection(useDb, 'newsNaver')

useCollection.find({'site':'Naver'}).count()

useCollection.aggregate([{ "$sort": { "_id": 1 } },
{
  "$group": {
    '_id': {'date':'$date', 'category': "$category", 'rank': "$rank" },
    "doc": { "$first": "$$ROOT" }
  }
},
{ "$replaceRoot": { "newRoot": "$doc" } },
{ "$out": "collection" }]);


useCollection.find({'site':'Naver'}).count()

useCollection2 = Use_Collection(useDb, 'comments')

useCollection2.aggregate([{ "$sort": { "_id": 1 } },
{
  "$group": {
    '_id': {'date':'$date', 'category': "$category", 'rank': "$rank" ,'comments':'$comments'},
    "doc": { "$first": "$$ROOT" }
  }
},
{ "$replaceRoot": { "newRoot": "$doc" } },
{ "$out": "collection" }]);
