==> personal_project/1_Preprocessing_for_Sentiment_Analysis.ipynb <==
# Preprocessin for Sentiment Analysis
> * [공공데이터포털](https://www.data.go.kr/) : 뉴스빅데이터 분석 정보
> * 수정일 2017.12.26 

# -*- coding: utf-8 -*-

import os
from glob import glob
import pandas as pd
import html
import itertools
import re

## 뉴스 빅데이터 현황으로부터 감정분석 문장 추출

path1 = './data/news/have_negative_positive/constitution/'
path2 = './data/news/have_negative_positive/household_debt/'
path3 = './data/news/have_negative_positive/olymphic/'
path4 = './data/news/have_negative_positive/'
pathList = [path1, path2, path3, path4]

def Get_CSVfileList(path):
    filelist = glob(path+'*.csv')
    filelist.extend(glob(path+'*.CSV'))
    return filelist

df_data=pd.DataFrame()
x = 0 
for path in pathList:
    fileList = Get_CSVfileList(path)
    for file in fileList:
        df = pd.read_csv(file,encoding='utf-8', engine='python')
        x += df.shape[0]
        if len(df_data) == 0:
            df_data = df[['감성분석', '긍정 문장', '부정 문장']]
        else:
            df_data = pd.concat([df_data, df[['감성분석', '긍정 문장', '부정 문장']]])
df_data.reset_index(drop=True, inplace=True)        
print (x)
print (df_data.shape)
df_data.head()

df_words = df_data['감성분석'].copy()
df_positive = df_data['긍정 문장'].copy()
df_negative = df_data['부정 문장'].copy()

* drop list
> 기자 & 이메일 주소만 있는 문장  
> Neutral 제거

## 긍정 문장 추출

### NaN & ' '제거

df_positive.dropna(inplace = True)
df_positive = df_positive.apply(html.unescape)
df_positive.reset_index(drop=True, inplace = True)
df_positive.drop(df_positive[df_positive.str.match('^ $')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

### #*****? 인 값 제거

df_positive.drop(df_positive[df_positive.str.match('^#[a-zA-Z0-9]+\?$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

### #*****! 인 값 제거

df_positive.drop(df_positive[df_positive.str.match('^#[a-zA-Z0-9]+\!$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

### 이메일만 있는 값 제거

df_positive.drop(df_positive[df_positive.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$')].index,inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

### 이메일과 기자명만 있는 값 제거

df_positive.drop(df_positive[df_positive.str.match('^[가-힣]{3,3}[. ]기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('^[가-힣]{3,3}기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*$')].index , inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('^[가-힣]{3,3}[. ]기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-. ]*$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+[가-힣]{3,3}기자')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+[. ][가-힣]{3,3}기자')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[가-힣 ]{3,4}기자$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[a-zA-Z0-9+-_. ]+[가-힣 ]{3,4}기자$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[a-zA-Z0-9+-_.]+@$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[a-zA-Z0-9+-_.]+$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[a-zA-Z0-9+-_.]+@([a-zA-Z0-9-]+\.[a-zA-Z0-9-.])*')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('.[가-힣]{3,4}[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*?')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.match('[가-힣+_.=]+?[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*?$')].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[df_positive.str.len() < 11].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.drop(df_positive[(df_positive.str.match('^[가-힣]{3,5}[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\| @]*?')) & (df_positive.str.len() < 100)].index, inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

print (df_positive[df_positive==''])

df_positive = df_positive.apply(html.unescape)

df_positive.drop_duplicates(inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)
df_positive = df_positive.apply(html.unescape)
df_positive.drop_duplicates(inplace = True)
df_positive.reset_index(drop=True, inplace = True)
print (df_positive.shape)

df_positive.to_csv('./data/sentiment_data/positive_sentiment_sentence_from_news.csv', index=False, header= False, encoding='utf-8')

## 부정 문장 추출

### NaN & ' '제거

df_negative.dropna(inplace = True)
df_negative = df_negative.apply(html.unescape)
df_negative.reset_index(drop=True, inplace = True)
df_negative.drop(df_negative[df_negative.str.match('^ $')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

### #*****? 인 값 제거

df_negative.drop(df_negative[df_negative.str.match('^#[a-zA-Z0-9]+\?$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

### #*****! 인 값 제거

df_negative.drop(df_negative[df_negative.str.match('^#[a-zA-Z0-9]+\!$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

### 이메일만 있는 값 제거

df_negative.drop(df_negative[df_negative.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$')].index,inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

### 이메일과 기자명만 있는 값 제거

df_negative.drop(df_negative[df_negative.str.match('^[가-힣]{3,3}[. ]기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('^[가-힣]{3,3}기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*$')].index , inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('^[가-힣]{3,3}[. ]기자[. ][a-zA-Z0-9+-_.]+@[a-zA-Z0-9-. ]*$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+[가-힣]{3,3}기자')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('^[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+[. ][가-힣]{3,3}기자')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[가-힣 ]{3,4}기자$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[a-zA-Z0-9+-_. ]+[가-힣 ]{3,4}기자$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[a-zA-Z0-9+-_.]+@$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[a-zA-Z0-9+-_.]+$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[a-zA-Z0-9+-_.]+@([a-zA-Z0-9-]+\.[a-zA-Z0-9-.])*')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('.[가-힣]{3,4}[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*?')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.match('[가-힣+_.=]+?[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\|]*?$')].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[df_negative.str.len() < 11].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.drop(df_negative[(df_negative.str.match('^[가-힣]{3,5}[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-.\| @]*?')) & (df_negative.str.len() < 100)].index, inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

print (df_negative[df_negative==''])

df_negative = df_negative.apply(html.unescape)

df_negative.drop_duplicates(inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)
df_negative = df_negative.apply(html.unescape)
df_negative.drop_duplicates(inplace = True)
df_negative.reset_index(drop=True, inplace = True)
print (df_negative.shape)

df_negative.to_csv('./data/sentiment_data/negative_sentiment_sentence_from_news.csv', index=False, header=False, encoding='utf-8')

df_positive = pd.concat([df_positive, pd.DataFrame(len(df_positive)*[1], columns=[1])],axis = 1)
df_positive.rename(columns={'긍정 문장':0},inplace = True)
df_negative = pd.concat([df_negative, pd.DataFrame(len(df_negative)*[0], columns=[1])],axis = 1)
df_negative.rename(columns={'부정 문장':0},inplace = True)


newsSentence = pd.concat([df_positive, df_negative], axis = 0)
newsSentence.reset_index(drop=True, inplace = True)
print (newsSentence.shape)
newsSentence.head()

## 특수문자 & 기자명 & 이메일 제거 및 50자 이내 제거

newsSentence[0] = newsSentence[0].str.replace('^[가-힣]{3,4}[//. ]?기자[//. ]',' ')

newsSentence[0] = newsSentence[0].str.replace('^[ ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]*',' ')

newsSentence[0] = newsSentence[0].str.replace('[가-힣]{3,4}[//. ]?기자[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+',' ')
print (newsSentence.shape)

newsSentence[0] = newsSentence[0].str.replace('[가-힣]{3,4}[//. ]?기자',' ')

newsPress = pd.read_csv('./data/newspress.txt',encoding='utf-8',header=None)
newsPress.head()

for idx in newsPress.index:
    newsSentence[0] = newsSentence[0].str.replace(newsPress.loc[idx][0],' ')

newsSentence[0] = newsSentence[0].str.replace('[☜\[\]☞※©◀◆◁◀◈○◌◎●◑◐☆★\|♠♡〈〉『」「』▲=⑨①②ⓒ③「」◎○~!@#$%^&*()_+|<>?:{}]', ' ')

newsSentence[0] = newsSentence[0].str.replace('[//. ]?[a-zA-Z0-9+-_.]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]*','')

newsSentence[0] = newsSentence[0].str.replace('[a-zA-Z0-9+-.]+@','')

newsSentence[0] = newsSentence[0].str.strip()

newsSentence.drop(newsSentence[newsSentence[0].str.len()<10].index, inplace = True)
newsSentence.reset_index(drop=True, inplace=True)
newsSentence.dropna(inplace=True)
print (newsSentence.shape)

newsSentence.drop(newsSentence[newsSentence[0].str.len()<50].index, inplace = True)
newsSentence.reset_index(drop=True, inplace=True)
newsSentence.dropna(inplace=True)
newsSentence.reset_index(drop=True, inplace=True)
newsSentence.drop_duplicates(inplace = True)
newsSentence.reset_index(drop=True, inplace=True)
print (newsSentence.shape)
newsSentence.to_csv('./data/sentiment_data/merged_sentiment_data_from_news_bigdata.txt',index=False, header=False)
newsSentence.head()

newsSentence = pd.read_csv('./data/sentiment_data/merged_sentiment_data_from_news_bigdata.txt',header=None,encoding='utf-8')
print (newsSentence.shape)
newsSentence.head()

## 네이버 영화 리뷰

def read_data(filename):
    with open(filename, 'r',encoding='utf-8') as f:
        data = [line.split('\t') for line in f.read().splitlines()]
        data = data[1:] # header
    return data
rating = pd.DataFrame(read_data('./data/sentiment_data/ratings.txt'))
rating.drop(0, axis = 1, inplace = True)
rating.rename(columns = {1:0, 2:1}, inplace = True)
print (rating.shape)
rating.head()

totalData = pd.concat([newsSentence, rating])
totalData.reset_index(drop=True, inplace = True)
totalData.dropna(inplace=True)
print (totalData.shape)
totalData.reset_index(drop=True, inplace = True)
totalData.drop(totalData[totalData[0] ==''].index, inplace = True)
print (totalData.shape)
totalData.reset_index(drop=True, inplace = True)
totalData.to_csv('./data/sentiment_data/merged_rawdata_for_sentiment_analysis.txt',index=False, header=False, encoding='utf-8')
print (totalData.shape)
totalData.head()

## 긍정 / 부정 단어 추출

df_words.dropna(inplace = True)
df_words = df_words.apply(html.unescape)
df_words.reset_index(drop = True, inplace = True)
df_words.drop(df_words[df_words.str.match('^ $')].index, inplace = True)
df_words.reset_index(drop = True, inplace = True)
df_words = df_words.apply(html.unescape)
print(df_words.shape)

df_words = df_words[df_words.str.match('NEGATIVE|POSITIVE')]
df_words.reset_index(drop = True, inplace = True)
print (df_words.shape)

negativeWords = df_words.str.findall('(NEGATIVE [가-힣]+)').str.join(' ')
negativeWords.dropna(inplace = True)
negativeWords.reset_index(drop=True, inplace = True)
positiveWords = df_words.str.findall('(POSITIVE [가-힣]+)').str.join(' ')
positiveWords.dropna(inplace = True)
positiveWords.reset_index(drop=True, inplace = True)
print (negativeWords.shape, positiveWords.shape)

positiveWords.drop(positiveWords[positiveWords==''].index, inplace = True)
positiveWords.reset_index(drop=True, inplace = True)
negativeWords.drop(negativeWords[negativeWords==''].index, inplace = True)
negativeWords.reset_index(drop=True, inplace = True)
print (negativeWords.shape, positiveWords.shape)
positiveWords = positiveWords.str.findall('([가-힣]+)')
negativeWords = negativeWords.str.findall('([가-힣]+)')

positiveWords = list(set(itertools.chain.from_iterable(positiveWords)))
negativeWords = list(set(itertools.chain.from_iterable(negativeWords)))
print (len(positiveWords), len(negativeWords))

positiveWords = pd.Series(positiveWords)
negativeWords = pd.Series(negativeWords)

positiveWords = pd.concat([positiveWords, pd.DataFrame(len(positiveWords) * [1], columns=[1])], axis = 1)
negativeWords = pd.concat([negativeWords, pd.DataFrame(len(negativeWords) * [0], columns=[1])], axis = 1)

positiveWords.to_csv('./data/sentiment_data/positive_sentiment_word_from_news.csv', index=False, header=False)
negativeWords.to_csv('./data/sentiment_data/negative_sentiment_word_from_news.csv', index=False, header=False)

rawdata = pd.concat([totalData, positiveWords, negativeWords])
print (rawdata.shape)

rawdata.to_csv('./data/sentiment_data/raw_data_for_sentiment.txt',index=False, header = False, encoding='utf-8')
