==> personal_project/4_Word2Vec_model_For_sentiment_analysis.ipynb <==
# Word2Vec model
> * Positive or Negative

import pickle
from tqdm import tqdm
tqdm.pandas(desc="progress-bar")

## Word2Vec model 생성을 위한 함수 정의

import multiprocessing
cores = multiprocessing.cpu_count()
#sg (int {1, 0}) – Defines the training algorithm. If 1, CBOW is used, otherwise, skip-gram is employed.
def Make_Word2Vec_Model(data, size, epoch, sg, window, cbow_mean, workers, negative, hs, tagger):
    from tqdm import tqdm
    tqdm.pandas(desc="progress-bar")
    from datetime import datetime
    from gensim.models import word2vec
    start = datetime.now()
    modelPath = './model/'
    modelName = 'word2vec_size-{}_epoch-{}_window-{}_negative-{}_hs-{}_sg-{}_cbow_mean-{}_by-{}.model'.format(
        size, epoch, window, negative, hs, sg, cbow_mean, tagger)
    modelName = modelPath+modelName
    print (modelName)
    w2v_model = word2vec.Word2Vec(size = size, sg = sg, cbow_mean = cbow_mean, negative = negative,
                                  hs = hs, window = window, workers = workers, iter=epoch)
    w2v_model.build_vocab(tqdm(data))
    w2v_model.train(tqdm(data), total_examples=w2v_model.corpus_count, epochs=w2v_model.iter) 
    end = datetime.now()
    w2v_model.save(modelName)
    print ("Total running time: ", end-start)
    return d2v_model

import os
from glob import glob
import pickle

from collections import namedtuple
from konlpy.tag import Mecab
from konlpy.utils import pprint
import numpy as np
from ckonlpy.tag import Twitter as ctwitter
mecab = Mecab()
ct = ctwitter()

from sklearn.linear_model import LogisticRegression
from gensim.models.doc2vec import TaggedDocument
from sklearn.model_selection import train_test_split

## Preprocessing

import pandas as pd

rawdata = pd.read_csv('./data/sentiment_data/raw_data_for_sentiment.txt',header=None,encoding='utf-8')
print (rawdata.shape)

rawdata.head()

model1 = Make_Word2Vec_Model(data = , size = 3000, epoch = 20, 
                   sg = 0, window = 10, workers = cores, negative = 7,
                    cbow_mean = 1, hs = 0, tagger = 'ct')

model2 = Make_Word2Vec_Model(data = , size = 3000, epoch = 20, 
                   sg = 0, window = 10, workers = cores, negative = 7,
                    cbow_mean = 0, hs = 0, tagger = 'ct')

model1 = Make_Word2Vec_Model(data = , size = 3000, epoch = 20, 
                   sg = 1, window = 10, workers = cores, negative = 7,
                    cbow_mean = 0, hs = 0, tagger = 'ct')

stopwords = open('./data/stopwordsList.txt',encoding='utf-8').readlines()
stopwords = list(map(lambda x: x.strip(), stopwords))

d2v_docs = list()
for idx in tqdm(rawdata.index):
    text = rawdata.loc[idx,0]
    pos = mecab.pos(text)
    pos = [x[0] for x in pos ]
    pos = [x for x in pos if not x in stopwords]
    label = rawdata.loc[idx,1]
    d2v_docs.append(TaggedDocument(pos, label))

train, test = train_test_split(d2v_docs, test_size = 0.1)

pickle.dump(d2v_docs,open('./data/pre_data/tagged_data/pre_data_d2v_docs_for_word2vec_sentiment_by_mecab.pickled','wb'))
pickle.dump(train,open('./data/pre_data/train_test_Data/pre_data_train_for_word2vec_sentiment_by_mecab.pickled','wb'))
pickle.dump(test,open('./data/pre_data/train_test_Data/pre_data_test_for_word2vec_sentiment_by_mecab.pickled','wb'))

train = pickle.load(open('./data/pre_data/train_test_Data/pre_data_train_for_word2vec_sentiment_by_mecab.pickled','rb'))
test = pickle.load(open('./data/pre_data/train_test_Data/pre_data_test_for_word2vec_sentiment_by_mecab.pickled','rb'))

x_train = [ x.words for x in tqdm(train)] 
y_train = [ x.tags for x in tqdm(train)] 
x_test = [ x.words for x in tqdm(test)] 
y_test = [ x.tags for x in tqdm(test)] 

w2v_model = word2vec.Word2Vec(sg=1, size=3000,window=8, workers = 6, iter = 20)
w2v_model.build_vocab(tqdm(x_train))
w2v_model.train(tqdm(x_train), total_examples=w2v_model.corpus_count, epochs=w2v_model.iter)

w2v_model.save('./model/w2v_model_size3000_window8_iter20_for_word2vec_sentiment_by_mecab.model')

w2v_model = word2vec.Word2Vec.load('./model/w2v_model_size3000_window8_iter20_for_word2vec_sentiment_by_mecab.model')

# importing bokeh library for interactive dataviz
import bokeh.plotting as bp
from bokeh.models import HoverTool, BoxSelectTool
from bokeh.plotting import figure, show, output_notebook

# defining the chart
output_notebook()
plot_tfidf = bp.figure(plot_width=700, plot_height=600, title="A map of word vectors",
    tools="pan,wheel_zoom,box_zoom,reset,hover,previewsave",
    x_axis_type=None, y_axis_type=None, min_border=1)

# getting a list of word vectors. limit to 10000. each is of 200 dimensions
word_vectors = [w2v_model[w] for w in list(w2v_model.wv.vocab.keys())[:10000]]

# dimensionality reduction. converting the vectors to 2d vectors
from sklearn.manifold import TSNE
tsne_model = TSNE(n_components=2, verbose=1, random_state=0)
tsne_w2v = tsne_model.fit_transform(word_vectors)
# putting everything in a dataframe
tsne_df = pd.DataFrame(tsne_w2v, columns=['x', 'y'])
tsne_df['words'] = list(w2v_model.wv.vocab.keys())[:10000]

# plotting. the corresponding word appears when you hover on the data point.
plot_tfidf.scatter(x='x', y='y', source=tsne_df)
hover = plot_tfidf.select(dict(type=HoverTool))
hover.tooltips={"word": "@words"}
show(plot_tfidf)

from sklearn.feature_extraction.text import TfidfVectorizer
print ('building tf-idf matrix ...')
vectorizer = TfidfVectorizer(analyzer=lambda x: x, min_df=10)
matrix = vectorizer.fit_transform(x_train)
tfidf = dict(zip(vectorizer.get_feature_names(), vectorizer.idf_))
print ('vocab size :', len(tfidf))

def buildWordVector(tokens, size):
    vec = np.zeros(size).reshape((1, size))
    count = 0.
    for word in tokens:
        try:
            vec += w2v_model[word].reshape((1, size)) * tfidf[word]
            count += 1.
        except KeyError: # handling the case where the token is not
                         # in the corpus. useful for testing.
            continue
    if count != 0:
        vec /= count
    return vec

from sklearn.preprocessing import scale
train_vecs_w2v = np.concatenate([buildWordVector(z, 3000) for z in tqdm(map(lambda x: x, x_train))])
train_vecs_w2v = scale(train_vecs_w2v)

test_vecs_w2v = np.concatenate([buildWordVector(z, 3000) for z in tqdm(map(lambda x: x, x_test))])
test_vecs_w2v = scale(test_vecs_w2v)

from keras.models import Model, Sequential
from keras.layers import Input, Dense, Embedding, embeddings, merge
model = Sequential()
model.add(Dense(32, activation='relu', input_dim=3000))
model.add(Dense(1, activation='sigmoid'))
model.compile(optimizer='rmsprop',
              loss='binary_crossentropy',
              metrics=['accuracy'])

model.fit(train_vecs_w2v, np.array(y_train), epochs=100, verbose=2)

score = model.evaluate(test_vecs_w2v, np.array(y_test), verbose=2)
print (score[1])

* 1500 : 0.835
* 3000 :  0.8513

model.save('./model/keras_model_size3000_window8_iter20_from_word2vec.model')
