==> personal_project/5_Train_Doc2Vec_model_For_ Sentiment_Analysis.ipynb <==
# Sentiment Analysis
> * Positive or Negative

import pickle
import html
import multiprocessing

import numpy as np
from tqdm import tqdm
tqdm.pandas(desc="progress-bar")
from collections import namedtuple
from gensim.models import doc2vec
from gensim.models.doc2vec import TaggedDocument

from konlpy.utils import pprint

from konlpy.tag import Mecab
from ckonlpy.tag import Twitter as ctwitter

from sklearn.model_selection import train_test_split

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestRegressor

## Doc2Vec

TaggedDocument = namedtuple('TaggedDocument', 'words tags')

#### twitter

train = pickle.load(open('./data/pre_data/train_test_Data/pre_by_ct_train.pickled','rb'))

model = doc2vec.Doc2Vec.load('./model/doc2vec_size2000_epoch15_window5_dbow_words_by_ct.model')

train_x = [model.infer_vector(doc.words) for doc in tqdm(train)]
train_y = [doc.tags[0] for doc in tqdm(train)]
pickle.dump(train_x,open('./data/pre_data/train_test_Data/train_x_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))
pickle.dump(train_y,open('./data/pre_data/train_test_Data/train_y_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del train
del train_x
del train_y

test = pickle.load(open('./data/pre_data/train_test_Data/pre_by_ct_test.pickled','rb'))

test_x = [model.infer_vector(doc.words) for doc in tqdm(test)]
test_y = [doc.tags[0] for doc in tqdm(test)]
pickle.dump(test_x,open('./data/pre_data/train_test_Data/test_x_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))
pickle.dump(test_y,open('./data/pre_data/train_test_Data/test_y_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del model

train_x = pickle.load(open('./data/pre_data/train_test_Data/train_x_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))
train_y = pickle.load(open('./data/pre_data/train_test_Data/train_y_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))

classifier = LogisticRegression(random_state=1234)
classifier.fit(train_x, train_y)

pickle.dump(classifier,open('./data/pre_data/classifier/classifier_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

classifier2 = RandomForestRegressor(random_state=1234)
classifier2.fit(train_x, train_y)

pickle.dump(classifier2,open('./data/pre_data/classifier/classifier_randomforest_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del train_x
del train_y

test_x = pickle.load(open('./data/pre_data/train_test_Data/test_x_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))
test_y = pickle.load(open('./data/pre_data/train_test_Data/test_y_by_ct_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))

print( classifier.score(test_x, test_y) )

print( classifier2.score(test_x, test_y) )

del test_x
del test_y
del classifier
del classifier2

* size 1000 : 0.640541518255
* size 2000 : 0.646232190561

#### mecab

train2 = pickle.load(open('./data/pre_data/train_test_Data/pre_by_mecab_train.pickled','rb'))

model2 = doc2vec.Doc2Vec.load('./model/doc2vec_size2000_epoch15_window5_dbow_words_by_mecab.model')

train2_x = [model2.infer_vector(doc.words) for doc in tqdm(train2)]
train2_y = [doc.tags[0] for doc in tqdm(train2)]
pickle.dump(train2_x,open('./data/pre_data/train_test_Data/train2_x_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))
pickle.dump(train2_y,open('./data/pre_data/train_test_Data/train2_y_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del train2
del train2_x
del train2_y

test2 = pickle.load(open('./data/pre_data/train_test_Data/pre_by_mecab_test.pickled','rb'))

test2_x = [model2.infer_vector(doc.words) for doc in tqdm(test2)]
test2_y = [doc.tags[0] for doc in tqdm(test2)]
pickle.dump(test2_x,open('./data/pre_data/train_test_Data/test2_x_by_mecab_from_doc2vec_siz2000_epoch15_window5_dbow_words.pickled','wb'))
pickle.dump(test2_y,open('./data/pre_data/train_test_Data/test2_y_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del model2

train2_x = pickle.load(open('./data/pre_data/train_test_Data/train2_x_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))
train2_y = pickle.load(open('./data/pre_data/train_test_Data/train2_y_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))

classifier_ = LogisticRegression(random_state=1234)
classifier_.fit(train2_x, train2_y)

pickle.dump(classifier_,open('./data/pre_data/classifier/classifier_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

classifier2_ = RandomForestRegressor(random_state=1234)
classifier2_.fit(train2_x, train2_y)

pickle.dump(classifier2_,open('./data/pre_data/classifier/classifier_randomforest_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','wb'))

del train2_x
del train2_y

test2_x = pickle.load(open('./data/pre_data/train_test_Data/test2_x_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))
test2_y = pickle.load(open('./data/pre_data/train_test_Data/test2_y_by_mecab_from_doc2vec_size2000_epoch15_window5_dbow_words.pickled','rb'))

print( classifier_.score(test2_x, test2_y) )

print( classifier2_.score(test2_x, test2_y) )

del test2_x
del test2_y
del classifier_
del classifier2_

* size 1000 : 0.64909839715
* size 2000 : 0.663053762244

from sklearn.preprocessing import scale
train_vecs_w2v = np.concatenate([z.reshape(1,-1) for z in tqdm(map(lambda x: x, train2_x))])
train_vecs_w2v = scale(train_vecs_w2v)

test_vecs_w2v = np.concatenate([z.reshape(1,-1) for z in tqdm(map(lambda x: x, test2_x))])
test_vecs_w2v = scale(test_vecs_w2v)

from keras.utils import np_utils
y_test2 = np_utils.to_categorical(test2_y,2)
y_train2 = np_utils.to_categorical(train2_y,2)

from keras_tqdm import TQDMCallback, TQDMNotebookCallback
from keras.models import Model, Sequential
from keras.regularizers import l2
from keras.optimizers import SGD
from keras.utils import np_utils
from keras.layers import Input, Dense, Embedding, embeddings, merge, Dropout, Activation
model = Sequential()
model.add(Dense(128, activation='relu', input_dim=2000))
model.add(Dense(64, activation='relu', kernel_regularizer=l2(0.001)))
model.add(Dropout(0.1))
model.add(Dense(32, activation='relu', kernel_regularizer=l2(0.001)))
model.add(Dropout(0.1))
model.add(Dense(2, activation='sigmoid'))
model.compile(#optimizer='rmsprop',
    optimizer='adadelta',
              #loss='binary_crossentropy',
    #optimizer=SGD(lr=0.2), 
    loss='binary_crossentropy',
    metrics=['accuracy'])

model.fit(train_vecs_w2v, y_train2,epochs=100, batch_size=50000, verbose=2)
score = model.evaluate(test_vecs_w2v, y_test2, verbose=2)
print (score[1])

mode2 = Sequential()
mode2.add(Dense(64, activation='relu', input_dim=2000))
mode2.add(Dropout(0.25))
mode2.add(Dense(32, activation='relu'))
mode2.add(Dropout(0.125))
mode2.add(Dense(2, activation='softmax'))
mode2.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
# Train model
mode2.fit(train_vecs_w2v, y_train2,epochs=100, batch_size=50000, verbose=2)
# Evaluate model
score, acc = model.evaluate(test_vecs_w2v, y_test2, verbose=2)


print('Score: %1.4f' % score)
print('Accuracy: %1.4f' % acc)

embedding_layer = Embedding(input_dim=len(list(model2.docvecs)),
                           output_dim = 2000)

from keras.layers import Activation
nn_model = Sequential()
nn_model.add(embedding_layer)
nn_model.add(LSTM(2))
nn_model.add(Dense(2))
nn_model.add(Activation('sigmoid'))
nn_model.compile(loss='binary_crossentropy',
                 optimizer='adam',
                 metrics=['accuracy'])
print('Train...')
nn_model.fit(train_vecs_w2v, y_train2,epochs=10, batch_size=25000)
# Evaluate model
score, acc = nn_model.evaluate(test_vecs_w2v, y_test2, verbose=2)
