==> personal_project/7-2_comments_Sentiment_analysis_Using_Doc2vec.ipynb <==
# 수집된 뉴스 기사 및 댓글에 대한 감정 분석
## * Doc2Vec 
* 데이터 
> 2017년 12월 1일부터 2018년 2월 1일까지 63일간 [네이버](http://www.naver.com)와 [다음](http://www.daum.net)의 랭킹뉴스와 뉴스의 댓글을 크롤링함.

import pickle
import html
import multiprocessing
from collections import namedtuple, OrderedDict
import re
import sys
import os
from glob import glob
import warnings

os.environ['KERAS_BACKEND']='tensorflow'

from numba import jit
import numpy as np
from tqdm import tqdm
tqdm.pandas(desc="progress-bar")
import pandas as pd

from gensim.models import doc2vec, KeyedVectors
from gensim.models.doc2vec import TaggedDocument

from konlpy.utils import pprint

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.metrics import confusion_matrix, classification_report, roc_curve,  accuracy_score
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.preprocessing import scale, MinMaxScaler
from sklearn.manifold import TSNE

import keras.backend.tensorflow_backend as K
from keras.preprocessing import sequence
from keras_tqdm import TQDMCallback, TQDMNotebookCallback
from keras.models import Model, Sequential
from keras.regularizers import l2
from keras.optimizers import SGD
from keras.utils import np_utils
from keras.layers import Input, Flatten, Dense, Embedding, embeddings, merge, Dropout, Activation,  LSTM, Bidirectional, SimpleRNN, GRU
from keras.layers.convolutional import Conv1D, Conv2D
from keras.layers.pooling import MaxPooling1D, GlobalMaxPooling1D
from keras.layers.normalization import BatchNormalization
from keras.layers.core import SpatialDropout1D
from keras.utils import np_utils
from tensorflow.python.client import device_lib
from keras.layers.merge import dot

import xgboost as xgb

import matplotlib.pyplot as plt

import bokeh.plotting as bp
from bokeh.models import HoverTool, BoxSelectTool
from bokeh.plotting import figure, show, output_notebook

import Database_Handler as dh
import Basic_Module as bm

from ckonlpy.tag import Twitter
from konlpy.tag import Mecab
ct = Twitter()
mecab = Mecab()

## Stopwords

stopwords = open('./data/stopwordsList.txt',encoding='utf-8').readlines()
stopwords = list(map(lambda x: x.strip(), stopwords))

## Load Data

#Naver
naverData = pickle.load(open('./data/pre_data/stastics/for_statistics_Naver_from_mongodb.pickled','rb'))
naverData = pd.DataFrame.from_dict(naverData, orient = 'index')
naverData.reset_index(inplace = True)
naverData.rename(columns = {'index' : 'id'}, inplace = True)
#Daum
daumData = pickle.load(open('./data/pre_data/stastics/for_statistics_daum_from_mongodb.pickled','rb'))
daumData = pd.DataFrame.from_dict(daumData, orient = 'index')
daumData.reset_index(inplace = True)
daumData.rename(columns = {'index' : 'id'}, inplace = True)

print ('Naver : {}'.format(naverData.shape))
print ('Daum : {}'.format(daumData.shape))

extNaverData = naverData.loc[:, ['id', 'title', 'date', 'press', 'rank']].copy()
extNaverData['site'] = pd.Series(['Naver']*extNaverData.shape[0])
extDaumData = daumData.loc[:, ['id', 'title', 'date', 'press', 'rank']].copy()
extDaumData['site'] = pd.Series(['daum']*extDaumData.shape[0])

if sys.platform =='darwin':
    loadModelPath = '/Volumes/disk1/model/'
    classifierPath = '/Volumes/disk1/data/pre_data/classifier/'
    #newsPath = '/Volumes/data/pre_data/news_sentiment/'
    newsPath = './data/pre_data/news_sentiment/'
elif sys.platform =='win32':
    loadModelPath = 'd:/model/'
    classifierPath = 'd:/data/pre_data/classifier/'
    newsPath = './data/pre_data/news_sentiment/'

## TaggedDocument For Doc2Vec

TaggedDocument = namedtuple('TaggedDocument', 'words tags sentiment')

## Doc2Vec Model

### Twitter

taggerName = 'ct'

#### News to tagged Document

if os.path.isfile('./data/pre_data/tagged_data/pre_data_daum_news_by_ct_for_doc2vec_sentiment_analysis.pickled'):
    taggedDaumData = pickle.load(open('./data/pre_data/tagged_data/pre_data_daum_news_by_ct_for_doc2vec_sentiment_analysis.pickled', 'rb'))
else:
    taggedDaumData = bm.MakeTaggedDataDAUM(daumData, TaggedDocument, ct, stopwords, 'daum')
    pickle.dump(taggedDaumData, open('./data/pre_data/tagged_data/pre_data_daum_news_by_ct_for_doc2vec_sentiment_analysis.pickled', 'wb'))

    
if os.path.isfile('./data/pre_data/tagged_data/pre_data_naver_news_by_ct_for_doc2vec_sentiment_analysis.pickled'):
    taggedNaverData = pickle.load(open('./data/pre_data/tagged_data/pre_data_naver_news_by_ct_for_doc2vec_sentiment_analysis.pickled', 'rb'))
else:
    taggedNaverData = bm.MakeTaggedDataDAUM(naverData, TaggedDocument, ct, stopwords, 'naver')
    pickle.dump(taggedNaverData, open('./data/pre_data/tagged_data/pre_data_naver_news_by_ct_for_doc2vec_sentiment_analysis.pickled', 'wb'))

#### Load Model

model1 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-10_negative-7_hs-0_dm-1_dm_concat-0_dm_mean-1_by-ct.model')
name1 = '-'.join(re.split('[\(\),\/]',str(model1)))+taggerName
model2 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-5_negative-7_hs-0_dm-1_dm_concat-1_dm_mean-0_by-ct.model')
name2 = '-'.join(re.split('[\(\),\/]',str(model2)))+taggerName
model3 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-None_negative-7_hs-0_dm-0_dm_concat-0_dm_mean-0_by-ct.model')
name3 = '-'.join(re.split('[\(\),\/]',str(model3)))+taggerName

#### tagged Document to Vector

# 다음
daum_news_by_m1_name = newsPath+'daum_news_vector_'+name1
daum_news_by_m2_name = newsPath+'daum_news_vector_'+name2
daum_news_by_m3_name = newsPath+'daum_news_vector_'+name3
if not os.path.isfile(daum_news_by_m1_name):
    daum_news_by_m1 = bm.Get_Infer_Vector(taggedDaumData, model1)
    pickle.dump(daum_news_by_m1, open(daum_news_by_m1_name, 'wb'))
    del daum_news_by_m1

if not os.path.isfile(daum_news_by_m2_name):
    daum_news_by_m2 = bm.Get_Infer_Vector(taggedDaumData, model2)
    pickle.dump(daum_news_by_m2, open(daum_news_by_m2_name, 'wb'))
    del daum_news_by_m2
    
if not os.path.isfile(daum_news_by_m3_name):
    daum_news_by_m3 = bm.Get_Infer_Vector(taggedDaumData, model3)
    pickle.dump(daum_news_by_m3, open(daum_news_by_m3_name, 'wb'))
    del daum_news_by_m3

# 네이버
naver_news_by_m1_name = newsPath+'naver_news_vector_'+name1
naver_news_by_m2_name = newsPath+'naver_news_vector_'+name2
naver_news_by_m3_name = newsPath+'naver_news_vector_'+name3
if not os.path.isfile(naver_news_by_m1_name):
    naver_news_by_m1 = bm.Get_Infer_Vector(taggedNaverData, model1)
    pickle.dump(naver_news_by_m1, open(naver_news_by_m1_name, 'wb'))
    del naver_news_by_m1

if not os.path.isfile(naver_news_by_m2_name):
    naver_news_by_m2 = bm.Get_Infer_Vector(taggedNaverData, model2)
    pickle.dump(naver_news_by_m2, open(naver_news_by_m2_name, 'wb'))
    del naver_news_by_m2
    
if not os.path.isfile(naver_news_by_m3_name):
    naver_news_by_m3 = bm.Get_Infer_Vector(taggedNaverData, model3)
    pickle.dump(naver_news_by_m3, open(naver_news_by_m3_name, 'wb'))
    del naver_news_by_m3

del taggedDaumData
del taggedNaverData

#### model1

classifierList = glob(classifierPath+'*'+name1)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m1 = pickle.load(open(daum_news_by_m1_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m1, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name1,index=None, encoding='utf-8')

%%time
naver_news_by_m1 = pickle.load(open(naver_news_by_m1_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m1, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name1,index=None, encoding='utf-8')

del daum_news_by_m1
del naver_news_by_m1
del daum_news_by_m1_name
del naver_news_by_m1_name
del name1
del model1
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver

#### model2

classifierList = glob(classifierPath+'*'+name2)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m2 = pickle.load(open(daum_news_by_m2_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m2, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name2,index=None, encoding='utf-8')

%%time
naver_news_by_m2 = pickle.load(open(naver_news_by_m2_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m2, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name2,index=None, encoding='utf-8')

del daum_news_by_m2
del naver_news_by_m2
del daum_news_by_m2_name
del naver_news_by_m2_name
del name2
del model2
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver

#### model3

classifierList = glob(classifierPath+'*'+name3)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m3 = pickle.load(open(daum_news_by_m3_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m3, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name3,index=None, encoding='utf-8')

%%time
naver_news_by_m3 = pickle.load(open(naver_news_by_m3_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m3, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name3,index=None, encoding='utf-8')

del daum_news_by_m3
del naver_news_by_m3
del daum_news_by_m3_name
del naver_news_by_m3_name
del name3
del model3
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver

### Mecab

taggerName = 'mecab'

#### News to tagged Document

if os.path.isfile('./data/pre_data/tagged_data/pre_data_daum_news_by_mecab_for_doc2vec_sentiment_analysis.pickled'):
    taggedDaumData = pickle.load(open('./data/pre_data/tagged_data/pre_data_daum_news_by_mecab_for_doc2vec_sentiment_analysis.pickled', 'rb'))
else:
    taggedDaumData = bm.MakeTaggedDataDAUM(daumData, TaggedDocument, mecab, stopwords, 'daum')
    pickle.dump(taggedDaumData, open('./data/pre_data/tagged_data/pre_data_daum_news_by_mecab_for_doc2vec_sentiment_analysis.pickled', 'wb'))

    
if os.path.isfile('./data/pre_data/tagged_data/pre_data_naver_news_by_mecab_for_doc2vec_sentiment_analysis.pickled'):
    taggedNaverData = pickle.load(open('./data/pre_data/tagged_data/pre_data_naver_news_by_mecab_for_doc2vec_sentiment_analysis.pickled', 'rb'))
else:
    taggedNaverData = bm.MakeTaggedDataDAUM(naverData, TaggedDocument, mecab, stopwords, 'naver')
    pickle.dump(taggedNaverData, open('./data/pre_data/tagged_data/pre_data_naver_news_by_mecab_for_doc2vec_sentiment_analysis.pickled', 'wb'))

#### Load Model

model1 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-10_negative-7_hs-0_dm-1_dm_concat-0_dm_mean-1_by-mecab.model')
name1 = '-'.join(re.split('[\(\),\/]',str(model1)))+taggerName
model2 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-5_negative-7_hs-0_dm-1_dm_concat-1_dm_mean-0_by-mecab.model')
name2 = '-'.join(re.split('[\(\),\/]',str(model2)))+taggerName
model3 = doc2vec.Doc2Vec.load(loadModelPath+'doc2vec_size-1000_epoch-20_window-None_negative-7_hs-0_dm-0_dm_concat-0_dm_mean-0_by-mecab.model')
name3 = '-'.join(re.split('[\(\),\/]',str(model3)))+taggerName

#### tagged Document to Vector

# 다음
daum_news_by_m1_name = newsPath+'daum_news_vector_'+name1
daum_news_by_m2_name = newsPath+'daum_news_vector_'+name2
daum_news_by_m3_name = newsPath+'daum_news_vector_'+name3
if not os.path.isfile(daum_news_by_m1_name):
    daum_news_by_m1 = bm.Get_Infer_Vector(taggedDaumData, model1)
    pickle.dump(daum_news_by_m1, open(daum_news_by_m1_name, 'wb'))
    del daum_news_by_m1

if not os.path.isfile(daum_news_by_m2_name):
    daum_news_by_m2 = bm.Get_Infer_Vector(taggedDaumData, model2)
    pickle.dump(daum_news_by_m2, open(daum_news_by_m2_name, 'wb'))
    del daum_news_by_m2
    
if not os.path.isfile(daum_news_by_m3_name):
    daum_news_by_m3 = bm.Get_Infer_Vector(taggedDaumData, model3)
    pickle.dump(daum_news_by_m3, open(daum_news_by_m3_name, 'wb'))
    del daum_news_by_m3

# 네이버
naver_news_by_m1_name = newsPath+'naver_news_vector_'+name1
naver_news_by_m2_name = newsPath+'naver_news_vector_'+name2
naver_news_by_m3_name = newsPath+'naver_news_vector_'+name3
if not os.path.isfile(naver_news_by_m1_name):
    naver_news_by_m1 = bm.Get_Infer_Vector(taggedNaverData, model1)
    pickle.dump(naver_news_by_m1, open(naver_news_by_m1_name, 'wb'))
    del naver_news_by_m1

if not os.path.isfile(naver_news_by_m2_name):
    naver_news_by_m2 = bm.Get_Infer_Vector(taggedNaverData, model2)
    pickle.dump(naver_news_by_m2, open(naver_news_by_m2_name, 'wb'))
    del naver_news_by_m2
    
if not os.path.isfile(naver_news_by_m3_name):
    naver_news_by_m3 = bm.Get_Infer_Vector(taggedNaverData, model3)
    pickle.dump(naver_news_by_m3, open(naver_news_by_m3_name, 'wb'))
    del naver_news_by_m3

del taggedDaumData
del taggedNaverData

#### model1

classifierList = glob(classifierPath+'*'+name1)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m1 = pickle.load(open(daum_news_by_m1_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m1, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name1,index=None, encoding='utf-8')

%%time
naver_news_by_m1 = pickle.load(open(naver_news_by_m1_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m1, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name1,index=None, encoding='utf-8')

del daum_news_by_m1
del naver_news_by_m1
del daum_news_by_m1_name
del naver_news_by_m1_name
del name1
del model1
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver

#### model2

classifierList = glob(classifierPath+'*'+name2)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m2 = pickle.load(open(daum_news_by_m2_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m2, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name2,index=None, encoding='utf-8')

%%time
naver_news_by_m2 = pickle.load(open(naver_news_by_m2_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m2, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name2,index=None, encoding='utf-8')

del daum_news_by_m2
del naver_news_by_m2
del daum_news_by_m2_name
del naver_news_by_m2_name
del name2
del model2
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver

#### model3

classifierList = glob(classifierPath+'*'+name3)

loadClassifierDict = dict(map(lambda x:bm.LoadClassifier(x), classifierList))

%%time
daum_news_by_m3 = pickle.load(open(daum_news_by_m3_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_daum = dict(map(lambda x: bm.PredictSentiment(daum_news_by_m3, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_daum = pd.DataFrame.from_dict(predictOutcome_daum)
predictOutcome_daum = extDaumData.merge(predictOutcome_daum,
                                   left_index = True, right_index = True)
predictOutcome_daum.to_csv('./outcome/outcome_news_sentiment_analysis_daum_'+name3,index=None, encoding='utf-8')

%%time
naver_news_by_m3 = pickle.load(open(naver_news_by_m3_name, 'rb'))
warnings.filterwarnings('ignore')
predictOutcome_naver = dict(map(lambda x: bm.PredictSentiment(naver_news_by_m3, x, loadClassifierDict[x]), loadClassifierDict))
predictOutcome_naver = pd.DataFrame.from_dict(predictOutcome_naver)
predictOutcome_naver = extNaverData.merge(predictOutcome_naver,
                                   left_index = True, right_index = True)
predictOutcome_naver.to_csv('./outcome/outcome_news_sentiment_analysis_naver_'+name3,index=None, encoding='utf-8')

del daum_news_by_m3
del naver_news_by_m3
del daum_news_by_m3_name
del naver_news_by_m3_name
del name3
del model3
del loadClassifierDict
del predictOutcome_daum
del predictOutcome_naver
