==> personal_project/7-2_News_Result_arrangement_for_sentiment_analysis.ipynb <==
# Result Arrangement of News Sentiment analysis

import pandas as pd
from glob import glob
import os
import pickle
import sys
import re
import dask.dataframe as dd
from collections import Counter
from sklearn.metrics import accuracy_score, classification_report, roc_curve
import numpy as np
import matplotlib.pyplot as plt

%matplotlib inline

def CounterOutcome(row):
    from collections import Counter
    id_ = row.name
    target = row[['LogisticRegression',
       'NeuralNetwork_1', 'NeuralNetwork_2', 'RandomForestClassifier',
       'SVC', 'XGBoost']]
    counter = dict(Counter(target))
    counter['id'] = id_
    return counter

plt.rcParams["font.family"]

import Basic_Module as bm

## Path

if sys.platform =='darwin':
    loadModelPath = '/Volumes/disk1/nmodel/'
elif sys.platform =='win32':
    loadModelPath = 'd:/model/'
classifierPath = './data/pre_data/classifier/'
outpath = './outcome/'

## Load Data

#Naver
naverData = pickle.load(open('./data/pre_data/stastics/for_statistics_Naver_from_mongodb.pickled','rb'))
naverData = pd.DataFrame.from_dict(naverData, orient = 'index')
naverData.reset_index(inplace = True)
naverData.rename(columns = {'index' : 'id'}, inplace = True)
naverData.set_index(keys='id', inplace = True)
#Daum
daumData = pickle.load(open('./data/pre_data/stastics/for_statistics_daum_from_mongodb.pickled','rb'))
daumData = pd.DataFrame.from_dict(daumData, orient = 'index')
daumData.reset_index(inplace = True)
daumData.rename(columns = {'index' : 'id'}, inplace = True)
daumData.set_index(keys = 'id', inplace = True)
print ('Naver : {}'.format(naverData.shape))
print ('Daum : {}'.format(daumData.shape))

naverData.head()

daumData.head()

outcomeDaumFileList = glob(outpath+'outcome_news_sentiment_analysis_daum'+'*')
outcomeNaverFileList = glob(outpath+'outcome_news_sentiment_analysis_naver'+'*')
print (len(outcomeDaumFileList), len(outcomeNaverFileList))

## Negative or Positive

outNaverDict = dict()
for idx, outcomeFile in enumerate(sorted(outcomeNaverFileList)):
    fileName = os.path.split(outcomeFile)[1]
    df = pd.read_csv(outcomeFile, index_col = 0, encoding='utf-8')
    out = df.apply(CounterOutcome, axis = 1)
    out = pd.DataFrame.from_records(out)
    out.rename({0 : 'negative', 1 : 'positive'}, axis = 1, inplace=True)
    out.set_index(keys='id', inplace = True)
    out.fillna(0, inplace=True)
    modelIs, tagger = bm.ExtractModelType(outcomeFile)
    mx = '{}_{}'.format(modelIs, tagger)
    print (mx)
    outNaverDict[mx] = out

outNaverDf = pd.DataFrame()
for idx,key in enumerate(outNaverDict):
    data = outNaverDict[key]
    print (key)
    if outNaverDf.shape == (0, 0):
        outNaverDf = data.copy()
    else:
        outNaverDf = outNaverDf.add(data)

outNaverDf.shape

mergedNaver = naverData.merge(outNaverDf, left_index=True, right_index= True)

mergedNaver['Decision'] = outNaverDf.idxmax(axis = 1)

mergedNaver.head()

mergedNaver.Decision.value_counts()

mergedNaver.to_csv('./outcome_for_News_sentiment_analysis/naver_news_sentiment_analysis.csv', index=True, header=True, encoding='utf-8')

outDaumDict = dict()
for idx, outcomeFile in enumerate(sorted(outcomeDaumFileList)):
    fileName = os.path.split(outcomeFile)[1]
    df = pd.read_csv(outcomeFile, index_col = 0, encoding='utf-8')
    out = df.apply(CounterOutcome, axis = 1)
    out = pd.DataFrame.from_records(out)
    out.rename({0 : 'negative', 1 : 'positive'}, axis = 1, inplace=True)
    out.set_index(keys='id', inplace = True)
    out.fillna(0, inplace=True)
    modelIs, tagger = bm.ExtractModelType(outcomeFile)
    mx = '{}_{}'.format(modelIs, tagger)
    print (mx)
    outDaumDict[mx] = out

outDaumDf = pd.DataFrame()
for idx,key in enumerate(outDaumDict):
    data = outDaumDict[key]
    print (key)
    if outDaumDf.shape == (0, 0):
        outDaumDf = data.copy()
    else:
        outDaumDf = outDaumDf.add(data)

outDaumDf.shape

mergedDaum = daumData.merge(outDaumDf, left_index=True, right_index= True)

mergedDaum['Decision'] = outDaumDf.idxmax(axis = 1)

mergedDaum.head()

mergedDaum.Decision.value_counts()

mergedDaum.to_csv('./outcome_for_News_sentiment_analysis/daum_news_sentiment_analysis.csv', index=True, header=True, encoding='utf-8')
